website-scraping-retrieval/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
altair==5.1.1
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cohere==4.24
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
dnspython==2.4.2
exceptiongroup==1.1.3
executing==1.2.0
Faker==19.6.0
fastavro==1.8.2
favicon==0.7.0
filelock==3.12.2
fonttools==4.42.1
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.35
greenlet==2.0.2
htbuilder==0.6.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipython==8.15.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyterlab-widgets==3.0.8
kaggle==1.5.16
kiwisolver==1.4.5
langchain==0.0.285
langsmith==0.0.35
llama-index==0.8.22
loguru==0.7.1
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
more-itertools==10.1.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==1.10.12
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
pymdown-extensions==10.3
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
safetensors==0.3.3
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
smmap==5.0.0
soupsieve==2.5
SQLAlchemy==2.0.20
st-annotated-text==4.0.1
stack-data==0.6.2
streamlit==1.26.0
streamlit-aggrid==0.3.4.post3
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.2
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-javascript==0.1.5
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.3
text-unidecode==1.3
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.1
trulens-eval==0.12.0
typing-inspect==0.8.0
typing_extensions>=4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==4.0.8
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
